// src/Model/data/colors.ts
export enum Colors {
  CITIES_BG = '#282c34',
  SOFT_WHITE = '#f2f2f2',
  PURPOLE = '#8e7cc3',
  GREY = '#808080',
  PARIS_BLUE = '#0055A4',
  LONDON_RED = '#CF081F',
  BERLIN_BLACK = '#323232',
  SPAIN_YELLOW = '#f1bf00',
  ITALY_GREEN = '#008C45',
  ISRAEL_BLUE = '#0038b8',
  JAPAN_RED = '#BC002D',
  CHINA_Red = '#EE1C25',
  KOREA_BLUE = '#0047A0',
  THAILAND_BLUE = '#2D2A4A',
  INDIA_ORANGE = '#FF671F',
  USA_BLUE = '#002868',
  MEXICO_GRENN = '#00933B',
}